Split contract.rs into logical sub-modules (queries, mutations)
The contract module has been restructured from a single file into a
directory-based module with separate files for:
- mod.rs: contract struct definition, helper functions, constants
- mutations.rs: state-changing functions (create, claim, complete,
  raise_dispute, update_bounty)
- queries.rs: read-only functions (get_bounty, get_contributor,
  get_bounty_count, get_bounties_by_status)

This improves maintainability by making the separation of concerns
explicit at the file system level.

Also removes committed test_snapshots/ from git tracking (they are
now gitignored).

Closes #309
Closes #303
Closes #307
Closes #308
