+++ src/contract/queries.rs
@@ -1,0 +1,31 @@
+use soroban_sdk::{contractimpl, Address, BytesN, Env, Symbol, Vec};
+
+use crate::storage;
+use crate::types::{Bounty, BountyMeta, Contributor};
+
+#[contractimpl]
+impl MergeMintContract {
+    pub fn get_bounty(env: Env, bounty_id: BytesN<32>) -> Option<Bounty> {
+        storage::get_bounty(&env, &bounty_id)
+    }
+
+    pub fn get_bounty_meta(env: Env, bounty_id: BytesN<32>) -> Option<BountyMeta> {
+        storage::get_bounty_meta(&env, &bounty_id)
+    }
+
+    pub fn get_contributor(env: Env, address: Address) -> Option<Contributor> {
+        storage::get_contributor(&env, &address)
+    }
+
+    pub fn get_bounty_count(env: Env) -> u64 {
+        storage::get_bounty_count(&env)
+    }
+
+    pub fn get_bounties_by_status(env: Env, status: Symbol) -> Vec<BytesN<32>> {
+        storage::get_bounties_by_status(&env, &status)
+    }
+
+    pub fn get_open_bounties(env: Env) -> Vec<BytesN<32>> {
+        storage::get_open_bounties(&env)
+    }
+}
